requirements.txt
cffi==1.17.1
contourpy==1.1.1
cryptography==44.0.1
cycler==0.12.1
fonttools==4.55.7
importlib_resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
packaging==24.2
pillow==10.4.0
pyasn1==0.6.1
pycparser==2.22
pyparsing==3.1.4
PySimpleGUI==5.0.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rsa==4.9
six==1.17.0
zipp==3.20.2
